requirements.txt
absl-py==0.9.0
adium-theme-ubuntu==0.3.4
alabaster==0.7.7
apache-beam==2.13.0
astor==0.8.1
astroid==1.4.4
atari-py==0.1.7
attrs==19.3.0
audioread==2.1.8
autograd==1.2
avro==1.9.0
Babel==1.3
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.4.1
bleach==3.1.0
bokeh==1.2.0
Bottleneck==1.2.1
bz2file==0.98
cachetools==3.1.1
ccsm==0.9.12.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
compizconfig-python==0.9.12.3
configparser==4.0.2
contextlib2==0.6.0.post1
crcmod==1.7
cvxopt==1.1.4
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
dill==0.2.9
docopt==0.6.2
docutils==0.12
dopamine-rl==2.0.3
entrypoints==0.3
enum34==1.1.6
fastavro==0.21.24
fasteners==0.15
Flask==1.0.3
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.3.0
gast==0.3.2
gevent==1.4.0
gin-config==0.1.4
google-api-core==1.11.1
google-api-python-client==1.7.9
google-apitools==0.5.28
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.6.1
google-cloud-bigtable==0.32.2
google-cloud-core==0.29.1
google-cloud-datastore==1.7.4
google-cloud-pubsub==0.39.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
googledatastore==7.0.2
greenlet==0.4.15
grpc-google-iam-v1==0.11.4
grpcio==1.26.0
gunicorn==19.9.0
gym==0.12.5
h5py==2.10.0
hdfs==2.5.2
html5lib==0.999
httplib2==0.12.0
idna==2.8
importlib-metadata==0.23
intervaltree==3.0.2
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jdcal==1.0
jedi==0.9.0
Jinja2==2.10.3
joblib==0.13.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==5.2.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kfac==0.1.4
kiwisolver==1.0.1
lazy-object-proxy==1.2.1
librosa==0.6.3
lifelines==0.14.0
llvmlite==0.29.0
logilab-common==1.1.0
lxml==3.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mercurial==3.7.3
mesh-tensorflow==0.0.5
mido==1.2.6
mir-eval==0.5
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==5.0.0
mpi4py==3.0.2
mpmath==1.1.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==1.8.1
nltk==3.4.1
nose==1.3.7
notebook==5.7.8
numba==0.44.0
numexpr==2.4.3
numpy==1.16.6
oauth2client==4.1.3
opencv-python==4.1.0.25
openpyxl==2.3.0
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
pathlib2==2.3.5
patsy==0.5.1
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plotly==4.6.0
prettytensor==0.7.4
progressbar2==3.39.3
prometheus-client==0.6.0
promise==2.2.1
prompt-toolkit==1.0.15
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.11.2
psutil==3.4.2
ptyprocess==0.6.0
py==1.4.31
pyarrow==0.13.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pydot==1.2.4
pyflakes==1.1.0
pyglet==1.3.2
Pygments==2.4.2
pygobject==3.20.0
pygtrie==2.3
pylint==1.5.2
pyparsing==2.0.3
pypng==0.0.19
pyrsistent==0.15.5
pytest==2.8.7
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2019.3
PyVCF==0.6.8
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
quadprog==0.1.7
requests==2.22.0
resampy==0.2.1
retrying==1.3.3
roman==2.0.0
rope==0.10.2
rsa==4.0
scandir==1.10.0
scikit-learn==0.20.4
scipy==1.2.3
scour==0.32
seaborn==0.9.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.14.0
sk-video==1.1.10
sklearn==0.0
sonnet==0.1.6
sortedcontainers==2.1.0
sox==1.3.7
Sphinx==1.3.6
sphinx-rtd-theme==0.1.9
spyder==2.3.8
stable-baselines==2.7.0
statsmodels==0.6.1
style==1.1.0
subprocess32==3.5.3
sympy==1.4
tables==3.2.2
tabulate==0.8.6
tensor2tensor==1.13.4
tensorboard==1.13.1
tensorflow-datasets==1.0.2
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-metadata==0.13.0
tensorflow-probability==0.6.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.3
trash-cli==0.12.9.14
typing==3.6.6
unity-lens-photos==1.0
update==0.0.1
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.4.2
wrapt==1.8.0
xlrd==0.9.4
xlwt==0.7.5
zipp==0.6.0
